- fix static analysis failure.   When calculating the next message queue head, the expression is promoted to type integer which causes a static analysis error when assigning it to the uint8_t head variable.   To fix this, explicitly cast the result of the arithmetic expression to a uint8_t.   This is safe against integer overflows due to the modulo operation in the expression.

+++ components/message-queue/implementation.c
@@ -333,7 +333,7 @@ bool
         {
             const uint16_t buffer_offset = mq->head * mq->message_size;
             memcopy((uint8_t *const)message, &mq->messages[buffer_offset], mq->message_size);
-            mq->head = (mq->head + 1) % mq->queue_length;
+            mq->head = (uint8_t)((mq->head + 1) % mq->queue_length);
             mq->available -= 1;
 
             if (mq->available == ({{message_queues.length}} - 1))
